# openapi/blobyard-core.openapi.yaml
openapi: 3.1.0
info:
  title: Blob Yard Core API
  version: 1.0.0
  description:
    The stable API for durable object storage and scoped access. Stored files do not expire when a
    transfer grant or public capability expires.
servers:
  - url: https://api.blobyard.com/v1
security:
  - bearerAuth: []
paths:
...
  /bootstrap/exchange:
    post:
      operationId: exchangeBootstrapToken
      x-blobyard-deployments:
        - self-hosted
      security: []
      responses:
        "200":
          description: Bootstrap authority exchanged once for a scoped local API token.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/ExchangeBootstrapTokenSuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - profiles
          - add
      x-blobyard-mcp:
        excluded: Bootstrap credentials must never enter model context.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./blobyard-shared.openapi.yaml#/components/schemas/ExchangeBootstrapTokenRequest
  /cli/whoami:
    get:
      operationId: whoAmI
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/WhoAmISuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - whoami
      x-blobyard-mcp:
        tool: blobyard_whoami
  /workspaces:
    get:
      operationId: listWorkspaces
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/ListWorkspacesSuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - workspaces
          - list
      x-blobyard-mcp:
        tool: blobyard_list_workspaces
      parameters:
        - in: query
          name: cursor
          required: false
          schema:
            $ref: ./blobyard-shared.openapi.yaml#/components/schemas/Cursor
    post:
      operationId: createWorkspace
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/CreateWorkspaceSuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - workspaces
          - create
      x-blobyard-mcp:
        tool: blobyard_create_workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./blobyard-shared.openapi.yaml#/components/schemas/CreateWorkspaceRequest
  /projects:
    get:
      operationId: listProjects
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/ListProjectsSuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - projects
          - list
      x-blobyard-mcp:
        tool: blobyard_list_projects
      parameters:
        - in: query
          name: cursor
          required: false
          schema:
            $ref: ./blobyard-shared.openapi.yaml#/components/schemas/Cursor
        - in: query
          name: workspace
          required: true
          schema:
            $ref: ./blobyard-shared.openapi.yaml#/components/schemas/Slug
    post:
      operationId: createProject
      responses:
        "200":
          description: Operation completed successfully.
          content:
            application/json:
              schema:
                $ref: ./blobyard-shared.openapi.yaml#/components/schemas/CreateProjectSuccessEnvelope
        default:
          $ref: ./blobyard-shared.openapi.yaml#/components/responses/ApiError
      x-blobyard-sdk: true
      x-blobyard-cli:
        command:
          - projects
          - create
      x-blobyard-mcp:
        tool: blobyard_create_project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ./blobyard-shared.openapi.yaml#/components/schemas/CreateProjectRequest
...
components:
  securitySchemes:
    bearerAuth:
      $ref: ./blobyard-shared.openapi.yaml#/components/securitySchemes/bearerAuth
    githubOidc:
      $ref: ./blobyard-shared.openapi.yaml#/components/securitySchemes/githubOidc
...